changed class to use delegates for compatibility
The mediaadmincontroller class for sonata is declared as a final class, therefore, no other classes can extend from it. The solution to this is to use delegates, where we create the same function signatures and simply pass it on to the sonata mediaadmincontroller class.

In order to preserve the functions under crudcontroller, our mediaadmincontroller class will extend from that instead.